📐 Architect: Refactor WeatherDashboard & Fix Lint
- Extracted `WeatherPanel`, `TidesPanel`, and `TasksPanel` from `WeatherDashboard` into separate component files, significantly reducing its token count and complexity.
- Created standard Architect documentation (`ai-index.md`, `ai-guidelines.md`, `.jules/architect-journal.md`).
- Fixed numerous pre-existing lint warnings in test files and explicitly suppressed ones that couldn't be fixed cleanly.
- Updated `npm run test:unit` and `npm run lint` pass flawlessly.
