+++ .jules/communication-board.md
@@ -1,0 +1,1 @@
+# Communication Board
