+++ src/components/__tests__/SettingsModal.test.tsx
@@ -5,7 +5,8 @@ import { describe, it, expect, vi, beforeEach } from 'vitest';
 
 // Mock framer-motion to avoid animation issues in tests
 vi.mock('framer-motion', () => {
-    // eslint-disable-next-line @typescript-eslint/no-require-imports
+
+    // eslint-disable-next-line @typescript-eslint/no-var-requires
     const React = require('react');
     const MOTION_PROPS = new Set(['initial', 'animate', 'exit', 'transition', 'whileHover', 'whileTap', 'variants', 'layout']);
     const MotionDiv = React.forwardRef(
